refactor(stable): import extractors types from package types barrel
Use ../../types for generated extractors model types instead of
importing directly from types.gen.ts.
